% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istack.r
\name{istack}
\alias{istack}
\title{Stack custom icons with group coloring}
\usage{
istack(D, var, group,
  icon = "https://teng-gao.github.io/images/person.png",
  icon_size = 0.03)
}
\arguments{
\item{D}{Data Frame}

\item{var}{categorical variable that is to be stacked}

\item{group}{grouping categorical variable that is to color the icons}

\item{icon}{link to any image icon on the internet. For coloring to work, has to be a PNG with transparent background}
}
\description{
Stack custom icons with group coloring
}
