% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istack.r
\name{istack}
\alias{istack}
\title{Stack custom icons with group coloring}
\usage{
istack(D, var, group, icon, size = 0.03, asp = 1)
}
\arguments{
\item{D}{Data Frame}

\item{var}{categorical variable that is to be stacked}

\item{group}{grouping categorical variable that is to color the icons}

\item{icon}{link to any image icon on the internet. For coloring to work, has to be a PNG with transparent background}

\item{size}{icon size}

\item{asp}{icon aspect ratio}
}
\value{
a ggplot object that can be modified downstream
}
\description{
Stack custom icons with group coloring
}
